mobile app refactoring
