+++ Products/mobile_app/lib/authenticator/jwt_authenticator.dart
@@ -1,0 +1,66 @@
+import 'dart:async';
+
+import 'package:flutter/widgets.dart';
+import 'package:mobile_app/authenticator/authentication_provider.dart';
+import 'package:mobile_app/authenticator.dart';
+import 'package:mobile_app/authenticator/errors.dart';
+import 'package:mobile_app/authenticator/jwt_authenticator/authentication_manager.dart';
+import 'package:mobile_app/authenticator/jwt_authenticator/tokens_manager.dart';
+
+/// Authenticator class to perform authentication of the user via a JWT provider.
+///
+/// This is a class that allows the management of JWT token with automatic refresh.
+final class JwtAuthenticator implements Authenticator {
+  final TokensManager _tokensManager;
+  final AuthenticationManager _authenticationManager;
+
+  JwtAuthenticator(final AuthenticationProvider provider)
+      : _tokensManager = TokensManager(provider),
+        _authenticationManager = AuthenticationManager(provider);
+
+  @override
+  Future<bool> authenticate(final BuildContext context) {
+    return _authenticationManager
+        .authenticate(context)
+        .then((code) => code != null);
+  }
+
+  @override
+  FutureOr<bool> authenticateIfAppropriate(final BuildContext context) async {
+    final token = await _tokensManager.tokens;
+
+    if (!context.mounted) {
+      debugPrint("Unmounted context");
+      return false;
+    }
+
+    debugPrint("hasToken: ${token != null}");
+
+    final code = await _authenticationManager.authenticateIfAppropriate(
+      context,
+      token != null,
+    );
+
+    return code != null;
+  }
+
+  /// #### Exceptions
+  /// Throws [TokenRefreshFailureException] in case of errors during token
+  /// refreshing.\
+  /// Throws [AuthenticationNotAskedException].
+  @override
+  Future<String?> authorization() async {
+    var tokens = await _tokensManager.tokens;
+
+    if (tokens == null) {
+      final code = _authenticationManager.code();
+      if (code == null) {
+        return null;
+      }
+
+      tokens = await _tokensManager.fromAuthorizationCode(code);
+    }
+
+    return "${tokens.tokenType} ${tokens.accessToken}";
+  }
+}
