+++ Products/mobile_app/lib/client/public_webserver_client.dart
@@ -1,0 +1,78 @@
+import 'dart:io';
+
+import 'package:dio/dio.dart';
+import 'package:flutter/foundation.dart';
+import 'package:mobile_app/authenticator.dart';
+import 'package:mobile_app/client.dart';
+import 'package:mobile_app/dtos/event.dart';
+import 'package:mobile_app/errors.dart';
+import 'package:result_dart/result_dart.dart';
+
+final class PublicWebServerClient implements Client {
+  static final defaultUri = Uri(
+    scheme: "http",
+    host: "192.168.1.15",
+    port: 8080,
+  );
+  static const _pathSegments = ['api', 'v3'];
+
+  final Uri uri;
+
+  final String apiKey;
+
+  final Dio _client;
+
+  factory PublicWebServerClient(String apiKey, [Uri? uri]) {
+    return PublicWebServerClient._(apiKey, uri ?? defaultUri);
+  }
+
+  PublicWebServerClient._(this.apiKey, this.uri)
+      : _client = Dio(BaseOptions(
+          baseUrl: uri.toString(),
+          queryParameters: {"api_key": apiKey},
+        ));
+
+  /// Throws [DioException] in case of networking errors.\
+  /// Throws [JsonValidationError] in case of invalid json.\
+  /// Throws [HttpStatusException] in case of status code
+  /// different from [HttpStatus.ok].\
+  /// Throws [TokenRefreshFailureException] in case of errors during token
+  /// refreshing.\
+  /// Throws [AuthenticationNotAskedException].
+  @override
+  getEventList() async {
+    final uri = this.uri.replace(
+      pathSegments: [..._pathSegments, 'events'],
+    );
+
+    // Optional authorization token.
+    String? token = await Authenticator().authorization();
+
+    debugPrint("Authorization: $token");
+
+    // TODO: convert DioException
+    final response = await _client.getUri(uri,
+        options: Options(headers: {
+          if (token != null) "Authorization": token,
+        }));
+
+    if (response.statusCode == HttpStatus.ok) {
+      final data = response.data;
+
+      if (data is! List) {
+        throw JsonValidationError(uri);
+      }
+
+      return data
+          .cast<Map<String, dynamic>>()
+          .map<Event>(Event.fromJson)
+          .toList(growable: false);
+    } else {
+      throw HttpStatusException(
+        uri,
+        response.statusCode ?? 0,
+        response.statusMessage ?? "Unknown error",
+      );
+    }
+  }
+}
