+++ Products/mobile_app/lib/dtos/category.dart
