+++ Products/mobile_app/lib/dtos/nullable_datetime_range.dart
