+++ Products/mobile_app/lib/errors/http_no_such_host_error.dart
@@ -1,0 +1,7 @@
+
+
+final class HttpNoSuchHostError extends Error {
+  final Uri uri;
+
+  HttpNoSuchHostError(this.uri);
+}
