+++ Products/mobile_app/lib/errors/http_status_exception.dart
@@ -1,0 +1,10 @@
+
+
+final class HttpStatusException implements Exception {
+  final Uri uri;
+
+  final int statusCode;
+  final String reasonPhrase;
+
+  HttpStatusException(this.uri, this.statusCode, this.reasonPhrase);
+}
