+++ Products/mobile_app/lib/errors/json_validation_error.dart
@@ -1,0 +1,7 @@
+
+
+final class JsonValidationError extends Error {
+  final Uri uri;
+
+  JsonValidationError(this.uri);
+}
