+++ Products/mobile_app/lib/routes/routes.gr.dart
@@ -1,0 +1,58 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// **************************************************************************
+// AutoRouterGenerator
+// **************************************************************************
+
+// ignore_for_file: type=lint
+// coverage:ignore-file
+
+// ignore_for_file: no_leading_underscores_for_library_prefixes
+import 'package:auto_route/auto_route.dart' as _i2;
+import 'package:flutter/material.dart' as _i3;
+import 'package:mobile_app/events/view/home_page.dart' as _i1;
+
+abstract class $BeeLiveRouter extends _i2.RootStackRouter {
+  $BeeLiveRouter({super.navigatorKey});
+
+  @override
+  final Map<String, _i2.PageFactory> pagesMap = {
+    HomeRoute.name: (routeData) {
+      final args =
+          routeData.argsAs<HomeRouteArgs>(orElse: () => const HomeRouteArgs());
+      return _i2.AutoRoutePage<dynamic>(
+        routeData: routeData,
+        child: _i1.HomePage(key: args.key),
+      );
+    }
+  };
+}
+
+/// generated route for
+/// [_i1.HomePage]
+class HomeRoute extends _i2.PageRouteInfo<HomeRouteArgs> {
+  HomeRoute({
+    _i3.Key? key,
+    List<_i2.PageRouteInfo>? children,
+  }) : super(
+          HomeRoute.name,
+          args: HomeRouteArgs(key: key),
+          initialChildren: children,
+        );
+
+  static const String name = 'HomeRoute';
+
+  static const _i2.PageInfo<HomeRouteArgs> page =
+      _i2.PageInfo<HomeRouteArgs>(name);
+}
+
+class HomeRouteArgs {
+  const HomeRouteArgs({this.key});
+
+  final _i3.Key? key;
+
+  @override
+  String toString() {
+    return 'HomeRouteArgs{key: $key}';
+  }
+}
