--- Products/mobile_app/lib/src/authenticator/authentication_provider/exceptions.dart
@@ -1,2 +1,0 @@
-
-
