--- Products/mobile_app/lib/src/features/login/view/login_page.dart
@@ -1,2 +1,0 @@
-
-
