+++ Products/mobile_app/pubspec.yaml
@@ -52,6 +52,7 @@ dependencies:
   flutter_map_cancellable_tile_provider: ^2.0.0
   draggable_bottom_sheet: ^1.0.2
   unixtime: ^1.0.0
+  result_dart: ^1.1.1
 
 dependency_overrides:
   http: ^1.2.1
